+++ src/space/core/entity/functions/entity_updates_personalise.rs
@@ -6,7 +6,7 @@ use crate::space::core::{
 
 pub fn personalise(
     updates_data: &mut HashMap<String, HashMap<String, EntityUpdateData>>,
-    player_handle: u32,
+    player_handle: u64,
     entity_updates_component: &EntityUpdates,
 ) {
     let mut to_be_removed_parameters = vec![];
